TiCpp: Rafactor, move each class to its own h/cpp
